Split denoising into local and slurm versions
